remove some rejects feature and remove some feature that will be added into meteor soon optimizes code
