+++ gradle.properties
@@ -1,17 +1,11 @@
-org.gradle.jvmargs = -Xmx4G
-
+org.gradle.jvmargs=-Xmx4G
 # Fabric Properties
-
-minecraft_version = 1.19.3
-yarn_version = 1.19.3+build.2
-loader_version = 0.14.11
-
+minecraft_version=1.19.3
+yarn_version=1.19.3+build.2
+loader_version=0.14.11
 # Mod Properties
-
-mod_version = 0.3.7
-maven_group = cally72jhb.addon
-archives_base_name = vector-addon
-
+mod_version=0.3.7
+maven_group=cally72jhb.addon
+archives_base_name=vector-addon
 # Meteor Client
-
-meteor_version = 0.5.2
+meteor_version=0.5.2
