+++ src/main/java/cally72jhb/addon/commands/commands/PlayerHeadCommand.java
@@ -1,19 +1,21 @@
 package cally72jhb.addon.commands.commands;
 
 import cally72jhb.addon.commands.arguments.PlayerNameArgumentType;
-import cally72jhb.addon.utils.ExecutorTask;
-import cally72jhb.addon.utils.ItemUtils;
-import cally72jhb.addon.utils.NetworkUtils;
-import com.google.gson.JsonObject;
 import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
+import meteordevelopment.meteorclient.systems.accounts.UuidToProfileResponse;
 import meteordevelopment.meteorclient.systems.commands.Command;
+import meteordevelopment.meteorclient.utils.network.Http;
+import meteordevelopment.meteorclient.utils.network.MeteorExecutor;
+import meteordevelopment.meteorclient.utils.player.FindItemResult;
+import meteordevelopment.meteorclient.utils.player.InvUtils;
 import net.minecraft.command.CommandSource;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.nbt.StringNbtReader;
+import net.minecraft.network.packet.c2s.play.CreativeInventoryActionC2SPacket;
 import net.minecraft.text.Text;
 import net.minecraft.util.math.MathHelper;
 
@@ -22,7 +24,8 @@ import java.util.Random;
 import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
 
 public class PlayerHeadCommand extends Command {
-    private final static SimpleCommandExceptionType NO_CREATIVE = new SimpleCommandExceptionType(Text.literal("You must be in creative mode to use this."));
+    private final static SimpleCommandExceptionType NOT_IN_CREATIVE = new SimpleCommandExceptionType(Text.literal("You must be in creative mode to use this."));
+    private final static SimpleCommandExceptionType NO_SPACE = new SimpleCommandExceptionType(Text.literal("No space in hotbar."));
 
     public PlayerHeadCommand() {
         super("player-head", "Gives you an player-head in creative.", "head", "skull");
@@ -53,34 +56,34 @@ public class PlayerHeadCommand extends Command {
 
     private void giveHead(String player, int amount) throws CommandSyntaxException {
         if (!mc.player.getAbilities().creativeMode) {
-            throw NO_CREATIVE.create();
+            throw NOT_IN_CREATIVE.create();
         } else if (player != null) {
-            ExecutorTask.execute(() -> {
+            MeteorExecutor.execute(() -> {
                 try {
                     ItemStack stack = Items.PLAYER_HEAD.getDefaultStack();
 
                     Random random = new Random(player.hashCode());
                     String id = "[I;" + random.nextInt() + "," + random.nextInt() + "," + random.nextInt() + "," + random.nextInt() + "]";
 
-                    JsonObject object = NetworkUtils.getJsonObject("https://api.mojang.com/users/profiles/minecraft/" + player);
-
-                    if (object != null) {
-                        JsonObject array = NetworkUtils.getJsonObject("https://sessionserver.mojang.com/session/minecraft/profile/" + object.get("id").getAsString());
-
-                        if (array != null) {
+                    APIResponse res = Http.get("https://api.mojang.com/users/profiles/minecraft/" + player).sendJson(APIResponse.class);
+                    if (res != null && res.name != null && res.id != null) {
+                        UuidToProfileResponse res2 = Http.get("https://sessionserver.mojang.com/session/minecraft/profile/" + res.id).sendJson(UuidToProfileResponse.class);
+                        if (res2 != null) {
                             if (amount > 1) stack.setCount(MathHelper.clamp(amount, 1, 64));
 
                             stack.setNbt(StringNbtReader.parse(
                                     "{SkullOwner:{Id:" + id + ",Properties:{textures:[{Value:\""
-                                            + array.get("properties").getAsJsonArray().get(0).getAsJsonObject().get("value").getAsString()
+                                            + res2.properties[0].value
                                             + "\"}]}}}")
                             );
+
+                            FindItemResult fir = InvUtils.find(ItemStack::isEmpty, 0, 8);
+                            if (!fir.found()) throw NO_SPACE.create();
+
+                            mc.getNetworkHandler().sendPacket(new CreativeInventoryActionC2SPacket(36 + fir.slot(), stack));
                         }
                     }
 
-                    if (!ItemUtils.insertCreativeStack(stack)) {
-                        error("Not enough space in your inventory.");
-                    }
                 } catch (CommandSyntaxException exception) {
                     exception.printStackTrace();
 
@@ -88,5 +91,9 @@ public class PlayerHeadCommand extends Command {
                 }
             });
         }
+    }
+
+    private static class APIResponse {
+        String name, id;
     }
 }
