--- src/main/java/cally72jhb/addon/mixin/GameRendererMixin.java
@@ -1,29 +1,0 @@
-package cally72jhb.addon.mixin;
-
-import cally72jhb.addon.commands.commands.ShaderCommand;
-import meteordevelopment.meteorclient.systems.commands.Commands;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gl.PostEffectProcessor;
-import net.minecraft.client.render.GameRenderer;
-import org.spongepowered.asm.mixin.Final;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(GameRenderer.class)
-public class GameRendererMixin {
-    @Shadow
-    @Final
-    MinecraftClient client;
-
-    @Inject(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/WorldRenderer;drawEntityOutlinesFramebuffer()V", ordinal = 0))
-    private void onRender(float tickDelta, long startTime, boolean tick, CallbackInfo info) {
-        PostEffectProcessor shader = Commands.get().get(ShaderCommand.class).getShader();
-        if (shader != null) {
-            shader.setupDimensions(client.getWindow().getFramebufferWidth(), client.getWindow().getFramebufferHeight());
-            shader.render(tickDelta);
-        }
-    }
-}
