+++ src/main/java/cally72jhb/addon/mixin/PlayerMoveC2SPacketAccessor.java
@@ -1,0 +1,21 @@
+package cally72jhb.addon.mixin;
+
+import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(PlayerMoveC2SPacket.class)
+public interface PlayerMoveC2SPacketAccessor {
+    @Mutable
+    @Accessor("x")
+    void setX(double x);
+
+    @Mutable
+    @Accessor("y")
+    void setY(double y);
+
+    @Mutable
+    @Accessor("z")
+    void setZ(double z);
+}
