+++ src/main/java/cally72jhb/addon/modules/movement/PacketDigits.java
@@ -1,11 +1,17 @@
 package cally72jhb.addon.modules.movement;
 
 import cally72jhb.addon.VectorAddon;
+import cally72jhb.addon.mixin.PlayerMoveC2SPacketAccessor;
+import cally72jhb.addon.mixin.VehicleMoveC2SPacketAccessor;
+import meteordevelopment.meteorclient.events.packets.PacketEvent;
 import meteordevelopment.meteorclient.settings.BoolSetting;
 import meteordevelopment.meteorclient.settings.IntSetting;
 import meteordevelopment.meteorclient.settings.Setting;
 import meteordevelopment.meteorclient.settings.SettingGroup;
 import meteordevelopment.meteorclient.systems.modules.Module;
+import meteordevelopment.orbit.EventHandler;
+import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
+import net.minecraft.network.packet.c2s.play.VehicleMoveC2SPacket;
 
 public class PacketDigits extends Module {
     private final SettingGroup sgGeneral = settings.getDefaultGroup();
@@ -34,27 +40,30 @@ public class PacketDigits extends Module {
             .build()
     );
 
-    // Constructor
-
     public PacketDigits() {
         super(VectorAddon.CATEGORY, "packet-digits", "Removes digits from your movement packets to make them smaller.");
     }
-
-    // Getter
-
-    public boolean shouldModifyFirstPacket() {
-        return firstPacket.get();
-    }
-
-    public boolean shouldModifyY() {
-        return modifyY.get();
-    }
-
-    // Utils
 
     public double round(double value) {
         int digit = (int) Math.pow(10, digits.get());
         double round = ((double) (Math.round(value * digit)) / digit);
         return Math.nextAfter(round, round + Math.signum(round));
+    }
+
+    @EventHandler
+    private void onSend(PacketEvent.Send event) {
+        if (firstPacket.get() && mc.player.age <= 20) {
+            if (event.packet instanceof PlayerMoveC2SPacket packet) {
+                if (packet.changesPosition()) {
+                    ((PlayerMoveC2SPacketAccessor) packet).setX(round(packet.getX(0)));
+                    ((PlayerMoveC2SPacketAccessor) packet).setY(modifyY.get() ? round(packet.getY(0)) : packet.getY(0));
+                    ((PlayerMoveC2SPacketAccessor) packet).setZ(round(packet.getZ(0)));
+                }
+            } else if (event.packet instanceof VehicleMoveC2SPacket packet) {
+                ((VehicleMoveC2SPacketAccessor) packet).setX(round(packet.getX()));
+                ((VehicleMoveC2SPacketAccessor) packet).setY(modifyY.get() ? round(packet.getY()) : packet.getY());
+                ((VehicleMoveC2SPacketAccessor) packet).setZ(round(packet.getZ()));
+            }
+        }
     }
 }
