+++ src/main/java/cally72jhb/addon/utils/Utils.java
@@ -48,7 +48,7 @@ public class Utils {
         double dx = forward * speed * cos + side * speed * sin;
         double dz = forward * speed * sin - side * speed * cos;
 
-        return new double[] { dx, dz };
+        return new double[]{dx, dz};
     }
 
     // Inventory
