+++ src/main/resources/vector-addon.mixins.json
@@ -4,10 +4,9 @@
   "package": "cally72jhb.addon.mixin",
   "compatibilityLevel": "JAVA_17",
   "client": [
-    "ClientConnectionMixin",
-    "ClientPlayNetworkHandlerMixin",
-    "GameRendererMixin",
-    "PlayerEntityMixin"
+    "PlayerEntityMixin",
+    "PlayerMoveC2SPacketAccessor",
+    "VehicleMoveC2SPacketAccessor"
   ],
   "injectors": {
     "defaultRequire": 1
